first pass at reducing warnings on Windows

+++ Opcodes/pinker.c
@@ -57,10 +57,10 @@ static int32_t instance_cnt = 0;    /* Is tis thread-safe? */
 
 #define F(cf,m,shift)   (0.0625f*cf*(2*((m)>>shift&1)-1))
 
-#define FA(n)   F(1.190566,n,0)+F(0.162580,n,1)+F(0.002208,n,2)+ \
-                F(0.025475,n,3)+F(-0.001522,n,4)+F(0.007322,n,5)-PINK_BIAS
-#define FB(n)   F(0.001774,n,0)+F(0.004529,n,1)+F(-0.001561,n,2)+ \
-                F(0.000776,n,3)+F(-0.000486,n,4)+F(0.002017,n,5)
+#define FA(n)   ((float)(F(1.190566,n,0)+F(0.162580,n,1)+F(0.002208,n,2)+ \
+                         F(0.025475,n,3)+F(-0.001522,n,4)+F(0.007322,n,5)-PINK_BIAS))
+#define FB(n)   ((float)F(0.001774,n,0)+F(0.004529,n,1)+F(-0.001561,n,2)+ \
+                 F(0.000776,n,3)+F(-0.000486,n,4)+F(0.002017,n,5))
 
 #define FA8(n)  FA(n),FA(n+1),FA(n+2),FA(n+3),FA(n+4),FA(n+5),FA(n+6),FA(n+7)
 #define FB8(n)  FB(n),FB(n+1),FB(n+2),FB(n+3),FB(n+4),FB(n+5),FB(n+6),FB(n+7)
